refactor: CategoryButton 컴포넌트 분리 (#22)
